+++ src/components/page/detail/Feed/FeedPanel.tsx
@@ -2,7 +2,8 @@ import { paths } from '@/__generated__/schema2';
 import { ampli } from '@/ampli';
 import { api } from '@/api';
 import { useMeetingQuery } from '@api/meeting/hook';
-import { useGetPostListInfiniteQuery, useMutationDeletePost, useMutationUpdateLike } from '@api/post/hooks';
+import { useGetPostListInfiniteQuery } from '@api/post/hooks';
+import { useMutationDeletePost, useMutationUpdateLike } from '@api/post/mutation';
 import { useUserProfileQuery } from '@api/user/hooks';
 import LikeButton from '@components/@common/button/LikeButton';
 import ContentBlocker from '@components/blocker/ContentBlocker';
